Add buttons to generate and clear grid
Put the grid generation loop in a function and link as a callback
to a generate button. The function clears existing grid first, then
takes in a user input number n to generate a n x n grid. The comparison
operator in the for loop converts the user input string to a number,
if valid.